% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_delete.R
\name{hist_delete}
\alias{hist_delete}
\title{Delete unintended source excerpts}
\usage{
hist_delete(filtered_sources, ID)
}
\arguments{
\item{filtered_sources}{Your current source dataframe}

\item{ID}{The text_ID of the source excerpts you want to remove}
}
\value{
The source dataframe without the source excerpts you deleted
}
\description{
This function deletes specific rows from a source dataframe created with hist_sources (text_output set on FALSE),
identified by their text_ID values.
}
\examples{
# Example source dataframe created with hist_sources
Xenophon_Corinthian <- hist_sources(author="Xenophon",context="Corinthian War", output_text = FALSE)
# Delete the source excerpts with IDs 5 and 14
hist_delete(Xenophon_Corinthian, ID = c(5, 14))
}
