% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_sources.R
\name{hist_sources}
\alias{hist_sources}
\title{Create a filtered source dataframe}
\usage{
hist_sources(
  author = NULL,
  context = NULL,
  event_type = NULL,
  event = NULL,
  work = NULL,
  citation = NULL,
  output_text = FALSE
)
}
\arguments{
\item{author}{The author to filter sources by (e.g. "Xenophon")}

\item{context}{The historical context (e.g. "Corinthian War")}

\item{event_type}{The event type (e.g. "battle")}

\item{event}{The exact event (e.g. "Battle1")}

\item{work}{The work from which you want to have source excerpts}

\item{citation}{The citation (e.g. "Citation1") of the respective source excerpts}

\item{output_text}{If TRUE the result text is provided formatted as output}
}
\value{
A filtered source dataframe (if output_text = FALSE)
}
\description{
This function filters the dataset the_sources according to the specified parameters.
If output_text = TRUE, the result text is provided formatted as output, ready to be incorporated into a source paper document
}
\examples{
# All source excerpts from Xenophon on the Corinthian War as dataframe
Xenophon_Corinthian <- hist_sources(author = "Xenophon", context = "Corinthian War")

# As source compilation
hist_sources(author = "Xenophon", context = "Corinthian War", output_text = TRUE)
}
