% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_text.R
\name{hist_text}
\alias{hist_text}
\title{Output a structured source compilation}
\usage{
hist_text(filtered_sources)
}
\arguments{
\item{filtered_sources}{The name of your current (filtered) source dataframe}
}
\value{
A structured source compilation of the sources of the respective source dataframe
}
\description{
This function converts a source dataframe that you created with hist_sources into a structured source compilation ready to be incorporated into a source paper document
}
\examples{
# Example source dataframe created with hist_sources
Xenophon_Corinthian <- hist_sources(author="Xenophon", context="Corinthian War", output_text = FALSE)
# Output the source-paper
hist_text(Xenophon_Corinthian)
}
